Elral/issue148 (#152)
* Split mobiflight.cpp into multiple files
Fixes #148
