+++ include/MFEEPROM.h
@@ -5,15 +5,13 @@
 /// \version 1.0 Initial release
 // Copyright (C) 2021
 
-#ifndef MFEEPROM_H
-#define MFEEPROM_H
+#pragma once
 
 class MFEEPROM
 {
 	  
   public:
     MFEEPROM();
-    void init(void);
     uint16_t get_length(void);
     void read_block(uint16_t addr, char data[], uint16_t len);
     void write_block (uint16_t addr, char data[], uint16_t len);
@@ -24,6 +22,3 @@ private:
     uint16_t eepromLength = 0;
 
 };
-
-
-#endif
