+++ include/config.h
@@ -1,0 +1,14 @@
+#pragma once
+
+void OnSetConfig();
+void OnResetConfig();
+void OnSaveConfig();
+void OnActivateConfig();
+void loadConfig();
+void OnGetConfig();
+void OnGetInfo();
+bool getStatusConfig();
+void generateSerial(bool);
+void OnGenNewSerial();
+void OnSetName();
+void _restoreName();
