+++ platformio.ini
@@ -31,6 +31,19 @@ build_flags =
 	-DMESSENGERBUFFERSIZE=96
 	-DMAXSTREAMBUFFERSIZE=96
 	-DDEFAULT_TIMEOUT=5000
+;	-DDEBUG2CMDMESSENGER=1
+	-I./src/MF_Analog
+	-I./src/MF_Button
+	-I./src/MF_Encoder
+	-I./src/MF_InputShifter
+	-I./src/MF_LCDDisplay
+	-I./src/MF_Output
+	-I./src/MF_OutputShifter
+	-I./src/MF_Segment
+	-I./src/MF_OutputShifter
+	-I./src/MF_Servo
+	-I./src/MF_OutputShifter
+	-I./src/MF_Stepper
 src_filter =
 	+<*>
 extra_scripts =
