+++ src/MF_Analog/Analog.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace Analog
+{
+void Add(uint8_t pin, char const *name, uint8_t sensitivity);
+void Clear();
+void read();
+void readAverage();
+}
