+++ src/MF_Encoder/Encoder.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace Encoder
+{
+void Add(uint8_t pin1, uint8_t pin2, uint8_t encoder_type, char const *name);
+void Clear();
+void read();
+}
