+++ src/MF_InputShifter/MFInputShifter.h
@@ -1,14 +1,6 @@
-// MFInputShifter.h
+#pragma once
 
-#ifndef MFInputShifter_h
-#define MFInputShifter_h
-
-#if ARDUINO >= 100
 #include <Arduino.h>
-#else
-#include <WProgram.h>
-#include <wiring.h>
-#endif
 
 // Maximum number of shifters allowed on an individual chain. While this is currently set to 4
 // there is no technical limit in the code for how many can be chained. It is constrained only
@@ -53,4 +45,3 @@ private:
   void clearLastState();
   static inputShifterEvent _inputHandler;
 };
-#endif
