+++ src/MF_LCDDisplay/LCDDisplay.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace LCDDisplay
+{
+void Add(uint8_t address, uint8_t cols, uint8_t lines);
+void Clear();
+void OnSet();
+}
