+++ src/MF_LCDDisplay/MFLCDDisplay.cpp
@@ -3,8 +3,6 @@
 // Copyright (C) 2013-2014
 
 #include "MFLCDDisplay.h"
-#include "allocateMem.h"
-#include "mobiflight.h"
 
 MFLCDDisplay::MFLCDDisplay()
 {
@@ -17,8 +15,8 @@ void MFLCDDisplay::display(const char *string)
     return;
   for (uint8_t line = 0; line != _lines; line++)
   {
-    _lcdDisplay->setCursor(0, line);
-    _lcdDisplay->writeString(&string[line*_cols], _cols);
+    _lcdDisplay.setCursor(0, line);
+    _lcdDisplay.writeString(&string[line*_cols], _cols);
   }
 }
 
@@ -27,18 +25,9 @@ void MFLCDDisplay::attach(byte address, byte cols, byte lines)
   _address = address;
   _cols = cols;
   _lines = lines;
-  if (!FitInMemory(sizeof(LiquidCrystal_I2C)))
-	{
-		// Error Message to Connector
-    cmdMessenger.sendCmdStart(kDebug);
-    cmdMessenger.sendCmdArg(F("LCD does not fit in Memory!"));
-    cmdMessenger.sendCmdEnd();
-		return;
-	}
-  _lcdDisplay = new (allocateMemory(sizeof(LiquidCrystal_I2C))) LiquidCrystal_I2C;
   _initialized = true;
-  _lcdDisplay->init((uint8_t)address, (uint8_t)cols, (uint8_t)lines);
-  _lcdDisplay->backlight();
+  _lcdDisplay.init((uint8_t)address, (uint8_t)cols, (uint8_t)lines);
+  _lcdDisplay.backlight();
   Wire.setClock(400000);
   test();
 }
@@ -53,9 +42,9 @@ void MFLCDDisplay::detach()
 void MFLCDDisplay::powerSavingMode(bool state)
 {
   if (state)
-    _lcdDisplay->noBacklight();
+    _lcdDisplay.noBacklight();
   else
-    _lcdDisplay->backlight();
+    _lcdDisplay.backlight();
 }
 
 void MFLCDDisplay::test()
@@ -63,7 +52,7 @@ void MFLCDDisplay::test()
   if (!_initialized)
     return;
   uint8_t preLines = 0;
-  _lcdDisplay->clear();
+  _lcdDisplay.clear();
 
   if (_lines > 2)
   {
@@ -76,7 +65,7 @@ void MFLCDDisplay::test()
     _printCentered("Happy Holidays!!", preLines++);
   }
 
-  _lcdDisplay->setCursor(0, 0);
+  _lcdDisplay.setCursor(0, 0);
 }
 
 void MFLCDDisplay::_printCentered(const char *str, uint8_t line)
@@ -90,10 +79,10 @@ void MFLCDDisplay::_printCentered(const char *str, uint8_t line)
     printChar = strlen(str);
   }
 
-  _lcdDisplay->setCursor(startCol, line);
+  _lcdDisplay.setCursor(startCol, line);
   
   for (uint8_t i = 0; i < printChar; i++)
   {
-    _lcdDisplay->write(str[i]);
+    _lcdDisplay.write(str[i]);
   }
 }
