+++ src/MF_Output/Output.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace Output
+{
+void Add(uint8_t pin);
+void Clear();
+void OnSet();
+void PowerSave(bool state);
+}
