+++ src/MF_OutputShifter/MFOutputShifter.cpp
