+++ src/MF_Servo/MFServo.h
@@ -6,17 +6,10 @@
 /// \author  Sebastian Moebius (mobiflight@moebiuz.de) DO NOT CONTACT THE AUTHOR DIRECTLY: USE THE LISTS
 // Copyright (C) 2013-2014 Sebastian Moebius
 
-#ifndef MFServo_h
-#define MFServo_h
+#pragma once
 
-#include <stdlib.h>
-#include <Servo.h>
-
-#if ARDUINO >= 100
 #include <Arduino.h>
-#else
-#include <WProgram.h>
-#endif
+#include <Servo.h>
 
 /////////////////////////////////////////////////////////////////////
 /// \class MFServo MFServo.h <MFServo.h>
@@ -36,9 +29,8 @@ private:
 		uint8_t _pin;
 		int			_mapRange[4];
     bool    _initialized;
-    Servo   *_servo;
+    Servo   _servo;
     long    _targetPos;
     long    _currentPos;    
     int     speed;
 };
-#endif
