+++ src/MF_Servo/Servos.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace Servos
+{
+void Add(int pin);
+void Clear();
+void OnSet();
+void update();
+}
